refactored tests and let percy run after the tests finished
