+++ tests/utils/generate_screenshot_test_report.py
@@ -14,7 +14,9 @@ def get_screenshot_test_base_folder():
 
     tox_env_name = os.getenv("TOX_ENV_NAME", "")
     dir_path = os.path.abspath(
-        os.path.join(os.path.dirname(__file__), "..", "test_screenshots", tox_env_name)
+        os.path.join(
+            os.path.dirname(__file__), "..", "..", "test_screenshots", tox_env_name
+        )
     )
     return dir_path
