+++ tox.ini
@@ -64,6 +64,5 @@ commands =
     {env:COMMAND:coverage} erase
     {env:COMMAND:coverage} run setup.py test
     {env:COMMAND:coverage} report
-#    py.test --cov=djangocms_equation --cov-config {toxinidir}/.coveragerc tests
